[SCB-1944]rename package name to avoid same class name in same package and cause coverage error
