+++ demo/demo-zeroconfig-schemadiscovery-registry/demo-zeroconfig-schemadiscovery-registry-server/src/main/java/org/apache/servicecomb/demo/zeroconfig/server/ServerApplication.java
@@ -15,10 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.demo.zeroconfig;
+package org.apache.servicecomb.demo.zeroconfig.server;
 
-import org.apache.servicecomb.demo.CategorizedTestCaseRunner;
+import java.util.concurrent.TimeUnit;
 import org.apache.servicecomb.demo.TestMgr;
+import org.apache.servicecomb.foundation.common.utils.BeanUtils;
 import org.apache.servicecomb.springboot2.starter.EnableServiceComb;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -26,22 +27,17 @@ import org.springframework.boot.builder.SpringApplicationBuilder;
 
 @SpringBootApplication
 @EnableServiceComb
-public class Application {
-
+public class ServerApplication {
   public static void main(final String[] args) throws Exception {
-    new SpringApplicationBuilder().sources(Application.class).web(WebApplicationType.SERVLET)
-        .build().run(args);
+    new SpringApplicationBuilder().sources(ServerApplication.class).web(WebApplicationType.SERVLET).build().run(args);
 
-    runTest();
+    SelfServiceInvoker invoker = BeanUtils.getBean("SelfServiceInvoker");
+    invoker.latch.await(10, TimeUnit.SECONDS);
+    TestMgr.check(invoker.result, "hello");
 
     TestMgr.summary();
     if (!TestMgr.errors().isEmpty()) {
-      throw new IllegalStateException("tests failed");
+      System.exit(1);
     }
-  }
-
-  private static void runTest() throws Exception {
-    CategorizedTestCaseRunner
-        .runCategorizedTestCase("demo-zeroconfig-schemadiscovery-registry-server");
   }
 }
