+++ demo/demo-zeroconfig-schemadiscovery-registry/demo-zeroconfig-schemadiscovery-registry-tests/pom.xml
@@ -29,7 +29,7 @@
   </parent>
 
   <properties>
-    <demo.main>org.apache.servicecomb.demo.zeroconfig.Application</demo.main>
+    <demo.main>org.apache.servicecomb.demo.zeroconfig.tests.Application</demo.main>
   </properties>
   <dependencies>
     <dependency>
