+++ demo/demo-zeroconfig-schemadiscovery-registry/demo-zeroconfig-schemadiscovery-registry-tests/src/main/java/org/apache/servicecomb/demo/zeroconfig/tests/Application.java
@@ -15,11 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.demo.zeroconfig;
+package org.apache.servicecomb.demo.zeroconfig.tests;
 
-import java.util.concurrent.TimeUnit;
+import org.apache.servicecomb.demo.CategorizedTestCaseRunner;
 import org.apache.servicecomb.demo.TestMgr;
-import org.apache.servicecomb.foundation.common.utils.BeanUtils;
 import org.apache.servicecomb.springboot2.starter.EnableServiceComb;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -27,17 +26,22 @@ import org.springframework.boot.builder.SpringApplicationBuilder;
 
 @SpringBootApplication
 @EnableServiceComb
-public class ServerApplication {
-  public static void main(final String[] args) throws Exception {
-    new SpringApplicationBuilder().sources(ServerApplication.class).web(WebApplicationType.SERVLET).build().run(args);
+public class Application {
 
-    SelfServiceInvoker invoker = BeanUtils.getBean("SelfServiceInvoker");
-    invoker.latch.await(10, TimeUnit.SECONDS);
-    TestMgr.check(invoker.result, "hello");
+  public static void main(final String[] args) throws Exception {
+    new SpringApplicationBuilder().sources(Application.class).web(WebApplicationType.SERVLET)
+        .build().run(args);
+
+    runTest();
 
     TestMgr.summary();
     if (!TestMgr.errors().isEmpty()) {
-      System.exit(1);
+      throw new IllegalStateException("tests failed");
     }
+  }
+
+  private static void runTest() throws Exception {
+    CategorizedTestCaseRunner
+        .runCategorizedTestCase("demo-zeroconfig-schemadiscovery-registry-server");
   }
 }
